major architectural change: 'context' is not stored permanently in widgets anymore, .validate() returns a new context which contains also information errors (instead of raising an exception).
In case of errors the user is responsible to push the context with error information to the widget so it can redisplay its value.

With this change widgets are mostly stateless (they only contain meaningful state while they are rendered). Still I don't want to make them completely thread-safe (as this would likely result in more complex code).